Create package duke
