--- src/main/data/tasklist.txt
@@ -1,3 +1,0 @@
-______________________________________________________________________________________
-1.[T][ ] run
-______________________________________________________________________________________
