+++ src/main/text-ui-test/ACTUAL.TXT
@@ -12,12 +12,12 @@ ________________________________________________________________________________
 
 Something went wrong: ./src/main/data/tasklist.txt (No such file or directory)
 ______________________________________________________________________________________
-Error: Invalid Task Index!
+Error: Invalid duke.Task Index!
 ______________________________________________________________________________________
 
 Something went wrong: ./src/main/data/tasklist.txt (No such file or directory)
 ______________________________________________________________________________________
-Error: Task is already marked as incomplete!
+Error: duke.Task is already marked as incomplete!
 ______________________________________________________________________________________
 
 Something went wrong: ./src/main/data/tasklist.txt (No such file or directory)
@@ -28,7 +28,7 @@ ________________________________________________________________________________
 
 Something went wrong: ./src/main/data/tasklist.txt (No such file or directory)
 ______________________________________________________________________________________
-Error: Task is already completed!
+Error: duke.Task is already completed!
 ______________________________________________________________________________________
 
 Something went wrong: ./src/main/data/tasklist.txt (No such file or directory)
