+++ src/main/text-ui-test/EXPECTED.TXT
@@ -14,11 +14,11 @@ You now have 1 task(s) in the list
 ______________________________________________________________________________________
 
 ______________________________________________________________________________________
-Error: Invalid Task Index!
+Error: Invalid duke.Task Index!
 ______________________________________________________________________________________
 
 ______________________________________________________________________________________
-Error: Task is already marked as incomplete!
+Error: duke.Task is already marked as incomplete!
 ______________________________________________________________________________________
 
 ______________________________________________________________________________________
@@ -27,7 +27,7 @@ Okay, I have marked this task as completed!
 ______________________________________________________________________________________
 
 ______________________________________________________________________________________
-Error: Task is already completed!
+Error: duke.Task is already completed!
 ______________________________________________________________________________________
 
 ______________________________________________________________________________________
