+++ src/main/text-ui-test/input.txt
@@ -4,8 +4,9 @@ unmark 0
 unmark 1
 mark 1
 mark 1
-deadline CS2103T Quiz /by Friday
-event project meeting /from Mon 2pm /to 4pm
+deadline CS2103T Quiz /by 2019-02-14
+event project meeting /from 2019-02-14 /to 2019-02-15
 list
+delete 1
 hello
 bye
